Create PSRC Facet Bar Charts

This function allows you to create facet bar charts.

Usage

create_facet_bar_chart(
  t,
  x,
  y,
  fill,
  g,
  moe = NULL,
  est = "percent",
  scales = "free",
  facet = 3,
  dec = 0,
  lpos = "above",
  color = "psrc_dark",
  title = NULL,
  subtitle = NULL,
  interactive = "no"
)

Arguments

t

A tibble or dataframe in long form for plotting

x

The name of the variable you want plotted on the X-Axis

y

The name of the variable you want plotted on the Y-Axis

fill

The name of the variable you want the fill color of the bars to be based on

g

The name of the variable to be the facets

moe

The name of the variable to be used for error bars, if desired - default to "NULL"

est

Type for the Y values - enter "percent", "currency" or "number", defaults to "percent"

scales

Value for axis in facets, either "fixed" or "free" - defaults to "free"

facet

Value for the number of columns in your facet - defaults to 3

dec

Number of decimal points in labels - defaults to 0

lpos

Position for the bar labels of "above" or "within" - defaults to "above"

color

Name of color palette to use - defaults to "psrc_dark"

title

Title to be used for chart, if desired - defaults to "NULL"

subtitle

Sub-title to be used for chart, if desired - defaults to "NULL"

interactive

Enable hover text and other interactive features - defaults to "no"

Value

facet bar chart that is either static or interactive depending on choice

Examples

library(dplyr)
# Read in the example data and filter to 2020 Population by Race by County
df <- read.csv(system.file('extdata', 'example_data.csv', package='psrcplot')) %>% 
      filter(Category=="Population by Race" & Year==2020) %>%
      filter(Geography!="Region" & Race!="Total")

# Create a facet chart for population by race using counties as the facet
my.chart <- create_facet_bar_chart(t=df, x="Race", y="share", 
                                   fill="Race", g="Geography", 
                                   facet=2, scales="fixed")