echart4r Bar Chart

echart4r Bar Chart

Usage

echart_bar_chart(
  t,
  x,
  y,
  est,
  fill,
  moe = NULL,
  pos = NULL,
  column_vs_bar = "column",
  ...
)

Arguments

t

A data frame in long form for plotting

x

The name of the category variable

y

The name of the variable with numeric values to plot

est

Select "number", "percent", or "currency" - defaults to "percent"

fill

The name of the variable you want the fill color of the bars or lines to be based on

moe

The name of the margin of error variable

pos

Determines if the bars are side-by-side (NULL) or stacked ("grp") - defaults to NULL

column_vs_bar

Select "column" or "bar" for chart orientation - defaults to "column"

...

additional arguments passed to generic_echart

Value

An interactive bar or column chart via echarts4r

Author

Christy Lam

Examples

library(tidyverse)
#> ── Attaching core tidyverse packages ──────────────────────── tidyverse 2.0.0 ──
#> ✔ forcats   1.0.0     ✔ readr     2.1.5
#> ✔ ggplot2   3.5.1     ✔ stringr   1.5.1
#> ✔ lubridate 1.9.3     ✔ tibble    3.2.1
#> ✔ purrr     1.0.2     ✔ tidyr     1.3.1
#> ── Conflicts ────────────────────────────────────────── tidyverse_conflicts() ──
#> ✖ readr::edition_get()   masks testthat::edition_get()
#> ✖ dplyr::filter()        masks psrcplot::filter(), stats::filter()
#> ✖ purrr::is_null()       masks testthat::is_null()
#> ✖ dplyr::lag()           masks stats::lag()
#> ✖ readr::local_edition() masks testthat::local_edition()
#> ✖ tidyr::matches()       masks dplyr::matches(), testthat::matches()
#> ℹ Use the conflicted package (<http://conflicted.r-lib.org/>) to force all conflicts to become errors

d <- mode_share_example_data |>
 filter(Year == 2020 & Category == 'Population by Race') |>
 mutate(Race = str_wrap(Race, 20))

color_palette <- c('#8dd3c7','#ffffb3','#bebada','#fb8072','#80b1d3')

echart_bar_chart(t = d,
                x = "Race",
                y = "share",
                fill = "Geography",
                color = color_palette,
                title = "Population by Race",
                subtitle = "In 2020",
                pos = NULL,
                column_vs_bar = "bar",
                est = "number")