Generic echart4r workhorse function Helper function to echart_bar_chart and echart_line_chart

Generic echart4r workhorse function Helper function to echart_bar_chart and echart_line_chart

Usage

generic_echart(
  df,
  category_var,
  fill = NULL,
  color = NULL,
  est = "percent",
  title = NULL,
  subtitle = NULL,
  legend = TRUE,
  egrid_bottom = "15%",
  egrid_left = "20%",
  legend_str_wrap = NULL,
  x_label_str_wrap = NULL,
  x_label_rotate = 0
)

Arguments

df

A data frame in long form for plotting

category_var

The name of the category variable

fill

The name of the variable you want the fill color of the bars or lines to be based on

color

A vector of colors or color palette

est

Select "number", "percent", or "currency" - defaults to "percent"

title

Chart title

subtitle

Chart subtitle

legend

TRUE or FALSE to display legend - defaults to TRUE

egrid_bottom

Bottom axis padding - defaults to "15%"

egrid_left

Left axis padding - defaults to "20%"

legend_str_wrap

Positive integer giving target line width to legend labels (in number of characters) - defaults to NULL

x_label_str_wrap

Positive integer giving target line width to x-axis labels (in number of characters) - defaults to NULL

x_label_rotate

Positive integer giving x-axis label rotation - defaults to 0

Value

Does not return a chart, called within echart_bar_chart and echart_line_chart